Cleanup storage of paxctld ansible variables
Converted current list structure:

```yaml
paxctld_configs:
  - binary: /usr/bin/python2.7
    flags: m
```

to the following structure:
```yaml
paxctld_configs:
  - /usr/bin/python2.7: m
```

This is really just a visual tweak for the human maintainer so the variables are a
little cleaner to look at and append to. I also removed some of the generic
defaults that I figured may not be applicable to all hosts. Next good step for
expansion would to make multiple layers of variables so we have sane defaults,
and hosts can append instead of having to override the entire `paxctld_configs`.