Adds default PaX flags from paxctld package
The paxctld package ships with a long list of default PaX flags that
this role was inappropriately clobbering. Adding those dist-maintained
flags back into the role logic, in a separate dict, so it can be
overridden with an empty dict if admins for some reason do not want
these flags.

Also sets the `paxctld_configs` dict to empty by default now, to provide
a single point of update for overrides.